Add code prevent duplicate data processed at consumer side

+++ src/main/java/com/agh/Payment/utils/VoucherPayment.java
@@ -12,7 +12,7 @@ public class VoucherPayment implements Payment{
 
 
     @Override
-    public PaymentResult pay(BigDecimal amount) {
+    public PaymentResult pay(double amount) {
         return new PaymentResult(
                 "Voucher payment successful",
                 PaymentStatus.SUCCESS,
